requirements.txt
asgiref==3.4.1
Django==3.2.6
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9.1
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.1
